requirements.txt
openai>=1.40.0
playwright>=1.45.0
pydantic>=2.7.0
jsonschema>=4.22.0
pyyaml>=6.0.1
python-dotenv>=1.0.1
aiohttp>=3.9.5
uvicorn>=0.30.1
fastapi>=0.111.0
rich>=13.7.1
structlog>=24.1.0
redis>=5.0.6
pytest>=8.2.2
pytest-asyncio>=0.23.7

